+++ src/timestamp.rs
@@ -1,5 +1,6 @@
+use rust_decimal::Decimal;
 #[derive(Clone)]
 pub struct Timestamp {
     pub hour: u8,
-    pub amount: f32,
+    pub amount: Decimal,
 }
